Workaround for BadStatusLine error
Workaround for https://github.com/mapillary/mapillary_tools/issues/113. Catch HTTPException due to which BadStatusLine error shows up. This way the thread where the exception appears is not interrupted and the file upload can be retried again until MAX_ATTEMPTS is reached. Don't know why the HTTPException is raised in the first place, but this way the sequence upload can be finished.